+++ src/ESP32FtpServer.h
@@ -93,7 +93,7 @@ private:
     boolean makePath(char *fullName, char *param);
     uint8_t getDateTime(uint16_t *pyear, uint8_t *pmonth, uint8_t *pday, uint8_t *phour, uint8_t *pminute, uint8_t *second);
     char* makeDateTimeStr(char *tstr, uint16_t date, uint16_t time);
-    int8_t readChar();
+    int16_t readChar();
 
     IPAddress dataIp;              // IP address of client for data
     WiFiClient client;
